requirements.txt
# Core dependencies
numpy>=1.20.0
scipy>=1.7.0
matplotlib>=3.4.0
Pillow>=8.0.0
python-dotenv>=0.19.0

# OpenAI API
openai>=1.0.0

# Gradio for the chat interface
gradio>=3.50.0

# Bruges for seismic modeling
bruges>=0.5.0